# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Application:
    properties:
      config:
        $ref: '#/definitions/models.JSONB'
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      oauth_clients:
        items:
          $ref: '#/definitions/models.OAuthClient'
        type: array
      protocol:
        description: oauth2, saml, ldap
        type: string
      saml_configs:
        items:
          $ref: '#/definitions/models.SAMLConfig'
        type: array
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  models.JSONB:
    additionalProperties: true
    type: object
...
  models.OAuthClient:
    properties:
      application_id:
        type: integer
      client_id:
        type: string
      created_at:
        type: string
      grant_types:
        items:
          type: string
        type: array
      id:
        type: integer
      redirect_uris:
        items:
          type: string
        type: array
      scopes:
        items:
          type: string
        type: array
    type: object
...
  models.SAMLConfig:
    properties:
      application_id:
        type: integer
      attribute_map:
        $ref: '#/definitions/models.JSONB'
      certificate:
        type: string
      created_at:
        type: string
      entity_id:
        type: string
      id:
        type: integer
      slo_url:
        type: string
      sso_url:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@openauth.local
    name: API Support
    url: https://github.com/hanyouqing/OpenAuth/issues
  description: OpenAuth - Open-source Identity and Access Management (IAM) platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/hanyouqing/OpenAuth
  title: OpenAuth API
  version: "1.0"
paths:
...
  /api-keys:
    get:
      description: Get list of API keys, optionally filtered by user or application
        (admin only)
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by application ID
        in: query
        name: application_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API key list
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      description: Create a new API key for user or application (admin only)
      parameters:
      - description: API key data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: API key created (key shown only once)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - api-keys
  /api-keys/{id}:
    delete:
      description: Delete API key by ID (admin only)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API key deleted
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - api-keys
    get:
      description: Get API key details by ID (admin only)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API key details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: API key not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get API key by ID
      tags:
      - api-keys
    put:
      consumes:
      - application/json
      description: Update API key information by ID (admin only)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: API key data to update
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: API key updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update API key
      tags:
      - api-keys
  /api-keys/{id}/revoke:
    post:
      description: Revoke (disable) an API key by ID (admin only)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API key revoked
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Revoke API key
      tags:
      - api-keys
  /applications:
    get:
      description: Get list of all applications (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Application list
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get application list
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: Create a new application (admin only)
      parameters:
      - description: Application data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Application'
      produces:
      - application/json
      responses:
        "200":
          description: Application created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create application
      tags:
      - applications
  /applications/{id}:
    delete:
      description: Delete application by ID (admin only)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Application deleted
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete application
      tags:
      - applications
    get:
      description: Get application details by ID (admin only)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Application details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Application not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get application by ID
      tags:
      - applications
    put:
      consumes:
      - application/json
      description: Update application information by ID (admin only)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: Application data to update
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Application updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update application
      tags:
      - applications
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"